FIX: String out-of-bounds error for references in strings
The problem lies in the fact that when deleting a string, the closing quote(s) get deleted first, and in every reference we used to shift the text range to the right by one place. The combination of the two resulted in an out-of-bounds error.

The fix simply checks if a PyStringLiteralExpression has balanced quotes before doing calculations on it

task: bb94902a-c7d2-494f-bb8b-f52ccc8318d5